components/notifications-badge: Extract "notifications" service
