+++ ember/app/templates/components/notifications-badge.hbs
@@ -1,1 +1,1 @@
-{{counterText}}
+{{notifications.counterText}}
